langchain1/requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20

# LLM Providers (choose what you need)
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-groq>=0.0.1

# Vector Stores & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Jupyter & Notebooks
ipykernel>=6.29.0
jupyter>=1.0.0
notebook>=7.0.0

# Visualization
graphviz>=0.20.1
matplotlib>=3.8.0

# Memory & Storage
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
aiohttp>=3.9.0

# Monitoring & Logging
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9
